# docs/api_contract.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "用户积分系统 API",
    "version": "1.0.0",
    "description": "积分获取、消费、冻结/解冻、冻结扣减、明细查询、余额查询接口契约。基于 docs/architecture.md 与 prisma/schema.prisma。"
  },
  "servers": [
    { "url": "/api/v1", "description": "API v1" }
  ],
  "paths": {
...
    "/points/freeze": {
      "post": {
        "tags": ["积分"],
        "summary": "积分冻结",
        "description": "冻结用户指定数量的积分，从可用余额转入冻结余额，冻结期间积分不可消费。冻结需设置超时时间，超时后由定时任务自动释放。支持幂等。",
        "operationId": "freezePoints",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FreezeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "冻结成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FreezeResult" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/AccountNotFound" },
          "409": { "$ref": "#/components/responses/IdempotentConflict" },
          "422": { "$ref": "#/components/responses/InsufficientBalance" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/points/balance": {
      "get": {
        "tags": ["积分"],
        "summary": "积分余额查询",
        "description": "查询用户当前可用积分余额、冻结余额及累计统计数据。",
        "operationId": "getBalance",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "maxLength": 64 },
            "description": "用户ID"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BalanceResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/AccountNotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FreezeRequest": {
        "type": "object",
        "required": ["user_id", "amount", "idempotent_key", "expires_at"],
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 64,
            "description": "用户标识"
          },
          "amount": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "description": "冻结积分数量"
          },
          "idempotent_key": {
            "type": "string",
            "maxLength": 128,
            "description": "幂等键，建议格式: freeze:{user_id}:{uuid_v4}"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "冻结超时时间，超时后定时任务自动解冻"
          },
          "remark": {
            "type": "string",
            "maxLength": 256,
            "description": "冻结原因（可选）"
          }
        }
      },
...
      "FreezeResult": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "example": "SUCCESS" },
          "message": { "type": "string", "example": "冻结成功" },
          "data": {
            "type": "object",
            "properties": {
              "freeze_id": { "type": "integer", "format": "int64", "description": "冻结记录ID" },
              "balance": { "type": "integer", "format": "int64", "description": "冻结后可用余额" },
              "frozen_balance": { "type": "integer", "format": "int64", "description": "冻结后冻结余额" },
              "expires_at": { "type": "string", "format": "date-time", "description": "冻结超时时间" }
            }
          }
        }
      },
...
      "BalanceResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "example": "SUCCESS" },
          "message": { "type": "string", "example": "查询成功" },
          "data": {
            "type": "object",
            "properties": {
              "user_id": { "type": "string" },
              "balance": { "type": "integer", "format": "int64", "description": "可用积分余额" },
              "frozen_balance": { "type": "integer", "format": "int64", "description": "冻结积分余额" },
              "total_earned": { "type": "integer", "format": "int64", "description": "累计获取积分" },
              "total_spent": { "type": "integer", "format": "int64", "description": "累计消费积分" },
              "version": { "type": "integer", "description": "乐观锁版本号" }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "机器可读错误码"
          },
          "message": {
            "type": "string",
            "description": "人类可读错误描述"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数校验失败",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "code": "POINTS_INVALID_PARAM",
              "message": "amount 必须大于 0"
            }
          }
        }
      },
      "AccountNotFound": {
        "description": "积分账户不存在",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "code": "POINTS_ACCOUNT_NOT_FOUND",
              "message": "用户积分账户不存在"
            }
          }
        }
      },
      "InsufficientBalance": {
        "description": "可用积分余额不足",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "code": "POINTS_INSUFFICIENT_BALANCE",
              "message": "可用积分余额不足"
            }
          }
        }
      },
      "IdempotentConflict": {
        "description": "幂等键冲突，重复请求",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "code": "POINTS_IDEMPOTENT_CONFLICT",
              "message": "重复请求，请使用原请求结果"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "系统内部错误",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "code": "POINTS_INTERNAL_ERROR",
              "message": "系统内部错误，请稍后重试"
            }
          }
        }
      }
    }
  },
...
}